fix: full-cycle review — git hardening, type safety, retry logic, recipe improvements
- Move validate_ref() to git/mod.rs, apply in all commit.rs functions
- StatusFilter::parse: replace unsafe &s[1..] with strip_prefix('!')
- HTTP retry: only retry on connect/timeout errors, not all failures
- Anonymizer: fix loose extension matching (ends_with("rs") → ".rs")
- Goal IDs: change goal_id/milestone_id from String to i64 for schema safety
- Goal error messages: standardize to "param is required for action 'name'"
- Memory recall: populate fact_type in semantic search results
- Session: fix limit cast to i64 in get_history_after_sync
- Setup wizard: fix step numbering gap (missing Step 3)
- Recipe: add implementation agent rules, sequence cargo update after code changes
- Update 78 dependencies to latest compatible versions

+++ crates/mira-server/src/db/session.rs
@@ -384,7 +384,7 @@ pub fn get_history_after_sync(
          ORDER BY id ASC
          LIMIT ?",
     )?;
-    let rows = stmt.query_map(params![session_id, after_id, limit], |row| {
+    let rows = stmt.query_map(params![session_id, after_id, limit as i64], |row| {
         Ok(ToolHistoryEntry {
             id: row.get(0)?,
             session_id: row.get(1)?,
